+++ onprem/ingest/__init__.py
@@ -1,2 +1,3 @@
 from .base import load_single_document, load_documents, process_folder, chunk_documents
 from .helpers import extract_files
+from .stores import VectorStoreFactory