+++ onprem/ingest/stores/__init__.py
@@ -1,5 +1,6 @@
 from .dense import DenseStore
 from .sparse import SparseStore
 from .dual import DualStore
+from .factory import VectorStoreFactory
 
-__all__ = ['DenseStore', 'SparseStore', 'DualStore']
+__all__ = ['DenseStore', 'SparseStore', 'DualStore', 'VectorStoreFactory']